backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
boto3==1.34.99
pydantic-settings==2.2.1
pydantic[email]==2.7.1
ultralytics>=8.3.0
Pillow==10.3.0
pyyaml==6.0.1
